detect and use _setmode() for Cygwin/MSYS, also used on Windows
use _setmode()

_setmode() is availble on all supported Windows platforms

It's also the recommended function to call:
https://learn.microsoft.com/cpp/c-runtime-library/reference/posix-setmode
https://learn.microsoft.com/cpp/c-runtime-library/reference/setmode

To make it complicated, it require `_O_BINARY` with an underscore.
`O_BINARY` is supported but bound to some conditions. Cygwin supports
_setmode, but always uses `O_BINARY` (no underscore). Cygwin maps
`setmode` to `_setmode` via a macro, so it's undetected by a function
check.

Either way on Windows `setmode(, O_BINARY)` was used with no issues
all along, so `_setmode(, O_BINARY)` should work just as well. If
this becomes a problem, logic can be added to use _O_BINARY on Windows.

_setmode() is availble on all supported Windows platforms 2