lib: WinCE is always _WIN32 too
This assumes there is not weird/hybrid WinCE build, which is not Win32,
while being WinCE. Something like Cygwin/MSYS, but for WinCE? Perhaps
the CeGCC build is like that, but that seems like a vaned off effort and
no binary that I found supports it.

Though it seems more likely this just an accident.